requirements.txt
# Core dependencies
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
pandas==2.1.3
numpy==1.24.3
sqlalchemy==2.0.23
sqlite3

# Web scraping
scrapy==2.11.0
lxml==4.9.3
fake-useragent==1.4.0

# Social media APIs
tweepy==4.14.0
facebook-sdk==3.1.0

# Notifications
smtplib
twilio==8.10.0
sendgrid==6.10.0

# Data analysis
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0

# Scheduling
schedule==1.2.0
apscheduler==3.10.4

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Database
sqlite3
psycopg2-binary==2.9.9

# Utilities
python-dateutil==2.8.2
pytz==2023.3
tqdm==4.66.1
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0
